Auto equip tools
This is optional and can be removed on demand. Need to fix the Anvil for the new auto equip. There are a few actions that needs to be added when this is used. Trees and stones are ready, so you may just check how it is done there, but i will fix it in the next days also.